Implementacion de servicio de imagen y documentacion

+++ backend/src/trabajo/trabajo.controller.ts
@@ -9,19 +9,30 @@ import {
   UploadedFile,
   UseInterceptors,
   BadRequestException,
+  NotFoundException,
 } from '@nestjs/common';
 import { TrabajoService } from './trabajo.service';
 import { CreateTrabajoDto } from './dto/create-trabajo.dto';
 import { UpdateTrabajoDto } from './dto/update-trabajo.dto';
-import { ApiTags, ApiConsumes, ApiBody } from '@nestjs/swagger';
+import {
+  ApiTags,
+  ApiConsumes,
+  ApiBody,
+  ApiOperation,
+  ApiResponse,
+} from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { diskStorage } from 'multer';
 import { extname } from 'path';
+import { ImageService } from '../imagenes/subir_image.service';
 
 @ApiTags('Trabajo')
 @Controller('trabajo')
 export class TrabajoController {
-  constructor(private readonly trabajoService: TrabajoService) {}
+  constructor(
+    private readonly trabajoService: TrabajoService,
+    private readonly imageService: ImageService,
+  ) {}
 
   @Post()
   @ApiConsumes('multipart/form-data')
@@ -63,16 +74,28 @@ export class TrabajoController {
       );
     }
 
-    const portadaTrabajo = `/uploads/portadasTrabajo/${file.filename}`;
-    return this.trabajoService.create(createTrabajoDto, portadaTrabajo);
+    const imagePath = await this.imageService.uploadImages(
+      [file],
+      'portadasTrabajo',
+    );
+
+    return this.trabajoService.create(createTrabajoDto, imagePath[0]);
   }
 
   @Get()
+  @ApiOperation({ summary: 'Obtiene todos los trabajos' })
+  @ApiResponse({
+    status: 200,
+    description: 'Lista de trabajos obtenida exitosamente.',
+  })
   findAll() {
     return this.trabajoService.findAll();
   }
 
   @Get(':id')
+  @ApiOperation({ summary: 'Obtiene un trabajo por su ID' })
+  @ApiResponse({ status: 200, description: 'Trabajo obtenido exitosamente.' })
+  @ApiResponse({ status: 404, description: 'Trabajo no encontrado.' })
   findOne(@Param('id') id: string) {
     return this.trabajoService.findOne(id);
   }
@@ -112,14 +135,38 @@ export class TrabajoController {
     @UploadedFile() file: Express.Multer.File,
     @Body() updateTrabajoDto: UpdateTrabajoDto,
   ) {
-    const portadaTrabajo = file
-      ? `/uploads/portadasTrabajo/${file.filename}`
-      : null;
-    return this.trabajoService.update(id, updateTrabajoDto, portadaTrabajo);
+    const existingTrabajo = await this.trabajoService.findOne(id);
+    if (!existingTrabajo) {
+      throw new NotFoundException(`Trabajo con id ${id} no encontrado`);
+    }
+
+    let newImagePath = null;
+
+    if (file) {
+      await this.imageService.deleteImages([existingTrabajo.portadaTrabajo]);
+
+      const newImagePaths = await this.imageService.uploadImages(
+        [file],
+        'portadasTrabajo',
+      );
+      newImagePath = newImagePaths[0];
+    }
+
+    return this.trabajoService.update(id, updateTrabajoDto, newImagePath);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @ApiOperation({ summary: 'Elimina un trabajo por su ID' })
+  @ApiResponse({ status: 200, description: 'Trabajo eliminado exitosamente.' })
+  @ApiResponse({ status: 404, description: 'Trabajo no encontrado.' })
+  async remove(@Param('id') id: string) {
+    const trabajo = await this.trabajoService.findOne(id);
+    if (!trabajo) {
+      throw new NotFoundException(`Trabajo con id ${id} no encontrado`);
+    }
+
+    await this.imageService.deleteImages([trabajo.portadaTrabajo]);
+
     return this.trabajoService.remove(id);
   }
 }
